+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in the last 3 major versions of the postgresql cookbook.
 
+## Unreleased
+
 ## 8.1.0 - *2020-12-09*
 
 - Fix potential password exposure in logs