+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 This file is used to list changes made in the last 3 major versions of the postgresql cookbook.
 
-## Unreleased
+## 8.2.0 - *2021-02-04*
 
-## Unreleased
+## 8.2.0 - *2021-02-04*
 
 - Makes psqlrc optional when invoking `psql_command_string`